[refactor] 전시회 등록 폼 common 폴더로 분리
